Move

+++ src/main/java/com/example/parking/service/ZoneServiceImpl.java
@@ -6,7 +6,9 @@ import com.example.parking.models.Zone;
 import com.example.parking.repository.LotRepository;
 import com.example.parking.repository.SpotRepository;
 import com.example.parking.repository.ZoneRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 @Service
@@ -28,28 +30,44 @@ public class ZoneServiceImpl implements ZoneService{
         }
         return zoneRepository.findAllByLotIdIs(id);
     }
+
     @Override
+    @Transactional
     public boolean deleteZone(Long id) throws NotFoundException {
         if(!zoneRepository.existsZoneByIdIs(id)){
             throw new NotFoundException("Zone not found");
         }
         //Delete spots in zone
-        spotRepository.deleteAllByZoneIdIs(id);
+        if(spotRepository.existsSpotByZoneIdIs(id)){
+            spotRepository.deleteAllByZoneIdIs(id);
+        }
         //Delete zones
         zoneRepository.deleteZoneByIdIs(id);
         return true;
     }
 
     @Override
-    public boolean addZone(Long lotId, Long length, Long width) {
+    public boolean addZone(Long lotId, String zoneName) {
         if(lotRepository.existsLotByIdIs(lotId)){
             try{
-                var newZone=new Zone(lotId,length,width);
+                var newZone=new Zone(lotId,zoneName);
                 zoneRepository.save(newZone);
                 return true;
             }
             catch(Exception e)
             {return false;}
+        }
+        return false;
+    }
+
+    @Override
+    public boolean editZone(Long zoneId, String zoneName) {
+        if(zoneRepository.existsZoneByIdIs(zoneId)){
+            int updatedRows = zoneRepository.updateZoneNameById(zoneId, zoneName);
+            if (updatedRows == 0) {
+                throw new EntityNotFoundException("Zone with id " + zoneId + " not found.");
+            }
+            return true;
         }
         return false;
     }
